Handle nested dataclasses
`from_dict()` now identifies fields of the current dataclass as another
dataclass. That dataclass is initialized with a `from_dict()` call
itself and fed back into the parent.

This leaves two primary inches to be developed.
1. A check that the inner dataclass has a `from_dict()` otherwise pass
the key/value as provided.
1. Currently using `globals()` to find and confirm what is a dataclass.
This feels hacky and should be replaced with a class decorater that
registers the dataclass.