Remove dependancy on importing future annotations
Prior to this commit it was required to have `__future__` imported with
the submodule `annotations`. This module normilized the type hints of
the classes.  Now we have a small regex search to parse out the class
name from the type strings which removes the prior dependancy.

Updated tests to include a set with `__future__`. There is test pollution
between the two suites. Future story to clean up how we handle multiple
class decorators with same names.

Updated example to remove use of `__future__`